ref(rust): rename snuba_options module to options
